requirements.txt
streamlit
requests
newspaper3k
spacy
dateparser
sentence-transformers
scikit-learn
openai
numpy
pandas
tqdm
python-dateutil
